for RTC#55592 CB-6212 - fix plugin warnings compiled under arm64-bit. Partial cherry-picks from cordova-plugin-contacts commit 0d7bf3980516b577deae0ff7f365b5c53116cbc1 and 910f80fae8fe79eb8342bcd737b663c2d5f62d78, cordova-plugin-geolocation commit 175f02ed6f19b4912eb345448d38d54c6c61d6e9, cordova-plugin-media-capture commit 007630da4866d867321008c6b8e1ebf57f05f259, cordova-plugin-inappbrowser commit fceea502a3f7f3308247a4bb2841b60a5d902c12, cordova-plugin-media commit 29fa80ba5dfa27f5be4e254442d72037cf3ff71e, cordova-plugin-file commit 53f72d85c93341d7828300a170f99bb6905c7edd, cordova-plugin-file-transfer commit 0f84287df03e7372eebe1d8620f8a78bb4900b4e, cordova-plugin-camera commit c9bab1f94c89e53c0ad8dbc77e1027eaa89527f1 and 91d6e10b29e492d8d72a2d5c759ebf7b876ecc3b, cordova-plugin-dialogs commit 9c7d4442f4f74cba573cff8a65deeb4437f21840, cordova-plugin-globalization commit d1b5e5bdebc954e149feecd745bdf9e4d5e2a668, cordova-plugin-network-information commit e0d3ae00b755540d163f7845ca728349f1c0ae4d

+++ CordovaLib/Classes/CDVCapture.m
@@ -419,7 +419,7 @@
         // NSLog(@"getFormatData: %@", [formatData description]);
     }
     if (bError) {
-        result = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageToErrorObject:errorCode];
+        result = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageToErrorObject:(int)errorCode];
     }
     if (result) {
         [self.commandDelegate sendPluginResult:result callbackId:callbackId];
@@ -784,7 +784,7 @@
 
     if (!self.pluginResult) {
         // return error
-        self.pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageToErrorObject:self.errorCode];
+        self.pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageToErrorObject:(int)self.errorCode];
     }
 
     self.avRecorder = nil;
